+++ packages/create/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/create
 
+## 0.12.37
+
+### Patch Changes
+
+-   Updated dependencies [ed6c364]
+    -   @sap-ux/preview-middleware@0.19.9
+    -   @sap-ux/app-config-writer@0.5.44
+
+## 0.12.36
+
+### Patch Changes
+
+-   Updated dependencies [5629071]
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.36
+
 ## 0.12.35
 
 ### Patch Changes